requirements.txt
# Core AI Framework
crewai

# LLM Integration
ollama

# Data validation
pydantic

# Environment variables
python-dotenv

# Testing
pytest

# YAML parsing
PyYAML

